requirements.txt
click>=8.1.0
pyyaml>=6.0
pandas>=2.0.0
jsonschema>=4.17.0

